이펙트 업데이트, 게임 루프 구현, 스킬 구현

+++ apiframeworktest/SkillGauge.h
@@ -26,7 +26,7 @@ private:
 
 private:
 	int m_max;
-	int m_val;
+	float m_val;
 
 	Image* m_backImg;
 	Image* m_fillImg;
